backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv>=1.0.0
crewai>=0.36.0
langchain-google-genai>=1.0.6
PyMuPDF==1.24.0
python-docx==1.1.0
